Update existing test scenes
